Defect 207 fixed. Show info dialog added

+++ Podcast/branches/s60withoutplayer/Engine/src/FeedEngine.cpp
@@ -591,47 +591,54 @@ void CFeedEngine::CompleteL(CHttpClient* /*aClient*/, TInt aError)
 		// the whole update chain will be broken
 		// change client state
 		iClientState = EIdle;
-		iActiveFeed->SetLastError(aError);
-		TTime time;
-		time.HomeTime();
-		iActiveFeed->SetLastUpdated(time);
-		
-		if( aError == KErrNone)
-			{			
-			TRAPD(parserErr, iParser->ParseFeedL(iUpdatingFeedFileName, iActiveFeed, iPodcastModel.SettingsEngine().MaxListItems()));
+		if(aError == KErrCancel)
+			{
+			iFeedsUpdating.Reset();
+			}
+		else
+			{
+			iActiveFeed->SetLastError(aError);
+			TTime time;
+			time.HomeTime();
+			iActiveFeed->SetLastUpdated(time);
 
-			if(parserErr)
-				{
-				// we do not need to any special action on this error.
-				iActiveFeed->SetLastError(parserErr);
-				DP1("CFeedEngine::Complete()\t Failed to parse feed. Leave with error code=%d", parserErr);
-				}
-			else
-				{
-				iPodcastModel.ShowEngine().DeleteOldShowsByFeed(iActiveFeed->Uid());
-				}
+			if( aError == KErrNone)
+				{			
+				TRAPD(parserErr, iParser->ParseFeedL(iUpdatingFeedFileName, iActiveFeed, iPodcastModel.SettingsEngine().MaxListItems()));
 
-			// delete the downloaded XML file as it is no longer needed
-			BaflUtils::DeleteFile(iPodcastModel.FsSession(),iUpdatingFeedFileName);			
-
-			// if the feed has specified a image url. download it if we dont already have it
-			if((iActiveFeed->ImageUrl().Length() > 0))
-				{
-				if ( (iActiveFeed->ImageFileName().Length() == 0) || 
-						(iActiveFeed->ImageFileName().Length() > 0 && 
-								!BaflUtils::FileExists(iPodcastModel.FsSession(), 
-										iActiveFeed->ImageFileName()) )
-				   )
+				if(parserErr)
 					{
-					TRAPD(error, GetFeedImageL(iActiveFeed));
-					if (error)
+					// we do not need to any special action on this error.
+					iActiveFeed->SetLastError(parserErr);
+					DP1("CFeedEngine::Complete()\t Failed to parse feed. Leave with error code=%d", parserErr);
+					}
+				else
+					{
+					iPodcastModel.ShowEngine().DeleteOldShowsByFeed(iActiveFeed->Uid());
+					}
+
+				// delete the downloaded XML file as it is no longer needed
+				BaflUtils::DeleteFile(iPodcastModel.FsSession(),iUpdatingFeedFileName);			
+
+				// if the feed has specified a image url. download it if we dont already have it
+				if((iActiveFeed->ImageUrl().Length() > 0))
+					{
+					if ( (iActiveFeed->ImageFileName().Length() == 0) || 
+							(iActiveFeed->ImageFileName().Length() > 0 && 
+									!BaflUtils::FileExists(iPodcastModel.FsSession(), 
+											iActiveFeed->ImageFileName()) )
+					)
 						{
-						// we have failed in a very early stage to fetch the image.
-						// continue with next Feed update	
-						iActiveFeed->SetLastError(parserErr);
-						iClientState = EIdle;							
-						}
-					}	
+						TRAPD(error, GetFeedImageL(iActiveFeed));
+						if (error)
+							{
+							// we have failed in a very early stage to fetch the image.
+							// continue with next Feed update	
+							iActiveFeed->SetLastError(parserErr);
+							iClientState = EIdle;							
+							}
+						}	
+					}
 				}
 			}
 		NotifyFeedUpdateComplete();
